+++ visitor-app/src/cb_admin/pages/RegisterU13.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Route, Switch } from 'react-router-dom';
 import { Grid } from 'react-flexbox-grid';
 import { assocPath } from 'ramda';
-import { RegisterMinorVisitor } from '../../shared/components/RegisterVisitor';
+import RegisterVisitor from '../../shared/components/RegisterVisitor';
 import NavHeader from '../../shared/components/NavHeader';
 import NotFound from '../../shared/components/NotFound';
 import { CommunityBusiness, Visitors, ErrorUtils } from '../../api';
@@ -33,7 +33,6 @@ export default class Main extends Component {
       cbOrgName: '',
       organisationId: null,
       cbLogoUrl: '',
-      ageCheck: false,
       // form validation for age
       signUpStatus: null,
     };
@@ -69,15 +68,15 @@ export default class Main extends Component {
   }
 
   validateForm = () => {
-    if (!this.state.phoneNumber && !this.state.email) {
+    if (!this.state.form.phoneNumber && !this.state.form.email) {
       return { errors: { email: 'You must supply a phone number or email address' } };
     }
 
-    if (!this.state.year) {
+    if (!this.state.form.year) {
       return { errors: { year: 'You must provide a year of birth' } };
     }
 
-    if (!this.state.ageCheck) {
+    if (!this.state.form.ageCheck) {
       return { errors: { ageCheck: 'You must confirm parental consent as been given' } };
     }
 
@@ -99,8 +98,10 @@ export default class Main extends Component {
       organisationId: this.state.organisationId,
     })
       .then((res) => {
-        this.setState({ qrCode: res.data.result.qrCode });
-        this.props.history.push('/visitor/signup/thankyou');
+        this.setState(
+          { qrCode: res.data.result.qrCode },
+          () => this.props.history.push('/admin/visitors/u-13/thankyou'),
+        );
       })
       .catch((err) => {
         if (ErrorUtils.errorStatusEquals(err, 400)) {
@@ -123,39 +124,39 @@ export default class Main extends Component {
     const { errors, cbOrgName, genders, cbLogoUrl, qrCode } = this.state;
 
     return (
-      <div className="row">
-        <Switch>
-          <Route exact path="/admin/visitors/u-13">
-            <Grid>
-              <NavHeader
-                leftTo="/admin"
-                leftContent="Back to dashboard"
-                centerContent="Sign Up Under 13's"
-              />
-              <RegisterMinorVisitor
-                handleChange={this.handleChange}
-                onSubmit={this.createVisitor}
-                cbName={cbOrgName}
-                years={BirthYear.u13OptionsList()}
-                genders={genders}
-                errors={errors}
-                status={this.state.signUpStatus}
-              />
-            </Grid>
-          </Route>
-
-          <Route path="/admin/visitors/u-13/thankyou">
-            <DisplayQrCode
-              cbLogoUrl={cbLogoUrl}
-              qrCode={qrCode}
-              nextURL="/admin"
-              onClickPrint={this.onClickPrint}
+      <Switch>
+        <Route exact path="/admin/visitors/u-13">
+          <Grid>
+            <NavHeader
+              leftTo="/admin"
+              leftContent="Back to dashboard"
+              centerContent="Sign Up Under 13's"
             />
-          </Route>
+            <RegisterVisitor
+              handleChange={this.handleChange}
+              onSubmit={this.createVisitor}
+              cbName={cbOrgName}
+              years={BirthYear.u13OptionsList()}
+              genders={genders}
+              errors={errors}
+              status={this.state.signUpStatus}
+              forMinor
+            />
+          </Grid>
+        </Route>
 
-          <Route exact path="/*" component={NotFound} />
-        </Switch>
-      </div>
+        <Route path="/admin/visitors/u-13/thankyou">
+          <DisplayQrCode
+            cbLogoUrl={cbLogoUrl}
+            qrCode={qrCode}
+            nextURL="/admin"
+            onClickPrint={this.onClickPrint}
+            forAdmin
+          />
+        </Route>
+
+        <Route exact path="/*" component={NotFound} />
+      </Switch>
     );
   }
 }
